Prepare the package to be installed via NPM
There are three main build targets:

  - dist/index.js

    A ES6 / CommonJS module that can be required or imported in projects 
that support modules.

  - dist/salary-calculator.js

    A script exposes global SalaryCalculator variable. Can be simply 
referenced in an HTML script tag. See website/custom.html

  - dist/index.html

    an HTML file that can be directly opened in the browser. Keep in 
mind that to work it needs other files in dist/ directory. If copying to 
another project, copy entire directory. See website/Makefile for example 
of that.

There is also a website/ directory containing a sample integration with 
outside projects. Later we can use it for:

  - Publishing stand alone website for Salary Calculator
  - Automated integration testing

Caveats:

  - I had to tweak Makefile a bit. Most notably the .installed file is 
now created before the npm ci is run. That's to avoid endless look (make 
-> npm ci -> make ...). I'm not happy about it. One problem is that if 
the npm ci command fails or is terminated, then make wrongly assumes 
that dependencies are installed. Another is that test, link and build 
are running twice after change in dependencies. For npm install there is 
--ignore-scripts flag that would be appropriate here, but ther 
apparently is no equivalent for npm ci.

  - The dist/index.html file assumes that script is at 
/salary-calculator.js (absolute path).  There may be a way to tweak 
parcel to make it relative.