Rewrite view logic without Elm Bootstrap
I aimed for maximal visual parity with previous implementation but there are
two unintended changes:

1. There is no animation when breakdown is showing / hiding

2. The "Okay, let's break it down..." toggle has additional margin

We could work it out, but IMO it actually looks better that way.

The tests won't pass. They need to be adjusted.

Improvements:

1. App became truly responsive - previously the breakdown formula would not
break on narrow viewports and overflow them along x axis.

2. We have greater control over markup

3. Code is more readable (IMHO), alas it probably can be further improved - I
worked in a hurry.

Other significant changes:

Apart from markup changes (using standard HTML elements + bootstrap classes),
      some changes were required to the Model type and update function. Model
      contains two additional fields:

1. breakdownVisible : Bool

2. activeField : Maybe Field

First one is hopefuly self explanatory. Active field reuses Filed type
previously only used to tag warnings. Now it also indicates which dropdown is
open. This way we have a type system guarantee that only one dropdown can be
open at any time.

There is a new set of Msg constructors. In general they are "lower level",
      meaning that they describe UI events like ...Clicked or ...ItemSelected.
      In longer run this should make the code more maintainable. Elm Bootstrap
      messages went away. In exchange we have two simpler messages:

1. BreakdownToggleClicked

2. DropdownButtonClicked Field

The second one takes advantage of the fact that we have 3 variable Fields
(Role, City and Tenure) and three corresponding dropdowns. These are the same
fields that are used to tag warnings and active field in the model. I'm not
sure if it's a good decision in long run. If the app evolves in a direction
where we use more (or less) dropdowns we may have to rethink it. But for now it
was very easy implementation so I went with it.

We don't need any subscriptions, however we may re-introduce it if we are going
to implement keyboard interactions (see below).

Minor changes:

I swapped the order of arguments for Warning constructor. In functional
programming it is good practice to have data as last argument (to enable easy
        currying). The message seems more data than Field tag, so it goes last.

I've extracted clearWarnings clearWarnings function from two lambdas we had
before.

Things I would like to improve:

1. Keyboard interactions: up and down arrow keys to navigate the dropdown
menus, esc to close them, etc.

2. Having smart positioning of the dropdowns taking edge of the viewport into
account. Maybe also the height of the dropdown, so it never overflows the
viewport.

3. ARIA attributes for accessability.